image.bbclass: make KERNEL_IMAGE & INITRD_IMAGE variable fixed
Before the value of those variables where calculated at runtime with the
content of the target root file system. But if the root file system
does not contain the files yet or no longer, this variable will be
empty, even if those files are already copyied to the deploy directory.

Setting this to fixed value allows to use them in every state of the
root file system.

Signed-off-by: Claudius Heine <ch@denx.de>