dpkg: warn when S equals WORKDIR
This patch does not change any logic, it just introduces a warning. If S
is not a directory inside WORKDIR do_deploy_deb and repo_clean will not
actually find any "*.deb" because they remain inside buildchroot.
Finding debs relative to BUILDROOT instead of S would require mounting
and locking. Intead of doing all of that document the small
limitation/assumption with a warning.

Signed-off-by: Henning Schild <henning.schild@siemens.com>