meta/classes: add image-account-extension class
This class allows to configure user and group accounts of the image.

Groups or users that should be configured/created are added into the
`GROUPS` or `USERS` variable.

The configuration itself is then added to each groups or users
`GROUP_<groupname>` or `USER_<username>` flags.

The flags available for groups are `gid` and `flags`. The `flags`
variable contains some additional options for the group. With this patch
only `system` is supported for groups, allowing to create groups with
`groupadd` with the `--system` parameter.

The flags available for users are `password`, `expire`, `inactive`,
`uid`, `gid`, `comment`, `home`, `shell`, `groups` and `flags`. The
additional flags for users are `no-create-home`, `create-home`, `system`
and `allow-empty-password`.

Signed-off-by: Claudius Heine <ch@denx.de>