image.bbclass: introduce KERNEL_FILE (defaults to vmlinuz)
Some Debian ports use a non-standard name for the symbolic link
to the Linux kernel: e.g. "vmlinux" for the mips port instead of
the usual "vmlinuz". The name of the symbolic link is needed for
Isar to copy the kernel image to tmp/deploy/images/<machine> (an
incorrect name causes the copy_boot_file task to fail).

Signed-off-by: Cedric Hombourger <Cedric_Hombourger@mentor.com>