image-postproc-extension: empty mechine-id
According to manpage of debian machine-id @
https://manpages.debian.org/buster/systemd/machine-id.5.en.html,
/etc/machine-id should be empty instead of removing.

Also according to above manpage, the dbus machine-id is copied to
/etc/machine-id if systemd machine-id is empty. However, the dbus
machine-id is also created during isar building, which stays the same
for every machine using the same image.

So here instead of removing, empty this file, and the mechine id will
be regenerated at the first boot. Also remove the dbus machine-id to
prevent copying.

Signed-off-by: Su Bao Cheng <baocheng.su@siemens.com>