meta: recipes-support shrink with preconfig instead of postproc
This adds two support packages to shrink an image and keep it small in
case apt/dpkg will be used on the system later on.
An alternative would be postprocess functions, they would be run-once
and if the system continues to receive package updates new files will
not be cleaned up.

Signed-off-by: Henning Schild <henning.schild@siemens.com>