image: Stop changing PF, tune WORKDIR, STAMP and IMAGE_FULLNAME directly
This removes the last fiddling with PF, addressing the issue - a common
image PN name for multiconfig targets while the results are actually
machine-specific - by tuning the critical paths directly. This are just
WORKDIR as well as STAMP and STAMPCLEAN.

For keeping the deployment filenames stable, set (and use consistently)
IMAGE_FULLNAME to the former PF value.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>