isar-bootstrap: Clean up and fix rebuild
There is no reason for using a lock anymore. All bootstraps for the same
distro/arch are detected by bitbake as identical targets in multiconfig
scenarios and are not run multiple times. The assumed original reason
for the lock, the bogus sharing of isar-bootstrap-host between Debian
and Raspbian, is fixed now.

In addition, remove the equally bogus re-execution test in
do_apt_config_prepare. This also fixes rebuilds of the isar-bootstrap
target in case some parameter has changed.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>