Detect false sharing of recipes
When a task is run more than once per build, this indicates a bug,
usually (though not only) related to incorrect multiconfig use. Such
bugs are hard to find in practice, so they could benefit from some
tooling.

This adds a very simple but effective check for such re-run scenarios
by putting a "once.<taskname>" stamp file into each workdir during a
build. If such a file already exists, an error is thrown. The workdirs
are purged after the parsing phase on the beginning of each build so
that no false positives are raised. The task stamp is also deleted after
failures so that no (serialized) retries can cause warnings.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>