Rework STAMP layout and cleaning
Follow OE more closely by aligning the stamps folder structure to
workdir. This may resolve false sharing of stamps between recipes of
common PF but different DISTRO/DISTRO_ARCH without the otherwise
required stamp-extra-info. That extra-info can now be removed.

As long as recipes mess with PF, we can't use OE's pattern as-is which
is PN-based.

Furthermore add proper STAMPCLEAN. This ensures that stamp files are
cleaned when a task is rebuilt. Not providing this pattern likely caused
a lot of failures when doing partial rebuilds.

As we are at it, switch do_clean to STAMPCLEAN as well.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>