u-boot-custom: Add control for u-boot-tools package build
Each custom u-boot currently generates also the u-boot-tools packages.
That can cause redundant builds in the best case when building multiple
targets with custom u-boots. In the worst case, an undeterministic
version is installed, depending one which u-boot with which version is
finished first.

Allow the user to define which recipe generates also the u-boot-tools
package so that only one is generated in a multi-target scenario. We
default to "no build" as a tools package different to the distro one is
rarely needed on the target.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>