meta-isar: Move bananapi to buster, remove bogus example
The bananapi machine config contained a "nice" anti-pattern that made
multiconfig explode: By tuning DISTRO_APT_PREFERENCES, it modified the
dependencies of isar-bootstrap-target for all other stretch-armhf
targets. And that nicely exploded when those were built in parallel.

Rather than trying to fix this pattern towards multiconfig
compatibility, which would have been rather complex, simply remove it.
We have buster now, and the buster kernel is what this target want. So
give it to him.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>