linux-custom: Control linux-libc-dev deployment manually
Deploying a version of linux-libc-dev that is different from the one
Debian uses easily causes problems. We already ran into those when doing
a downgrade, but we can also create deadlocks when doing an update. The
latter happens in common cross-build scenarios when pushing a new
version for the target arch but not providing one for the builder.

Avoid such troubles my making the package deployment opt-in. In most
cases, we will not depend on such an update because we will rarely
exploit new kernel API in userspace packages.

We can revert this behavior once we support building packages for both
target and host.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>