buildchroot: Factor out common packages
No need to maintain this list twice.

At this chance, make adduser a common package. be291cd991bd added it
only to the target buildchroot, while it is used by both. That fact
indicates that we only list it here for future cases when it is no
longer implicitly installed.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>