meta/classes: Generate bill of materials from image
To create products it is necessary to have a list
of used packages for clearance and security monitoring.
To get a simple list of packages use dpkg-query and generate
a list with the following pattern:

source name| source version | binary package name | binary version

All rootfs generate the list by default. Currently the
following lists are generated:
- buildchroot-${DISTRO}-${ARCH}.manifest
- ${IMAGE}-${DISTRO}-${ARCH}.manifest
- optional: sdkchroot-${DISTRO}-${ARCH}.manifest

Remove the feature with:
ROOTFS_FEATURES_remove = "generate-manifest"

Signed-off-by: Quirin Gylstorff <quirin.gylstorff@siemens.com>