bitbake: update to version 1.44.0
Found that the gitsm fetcher we ship is rather out-dated compared to
upstream and was failing to pull git sub-modules from repositories
such as azure-umqtt-c.

It shall be noted that bitbake has renamed its multiconfig targets:
the "multiconfig:" prefix was changed to "mc:". Updated our scripts
as well as documentation.

    Origin: https://github.com/openembedded/bitbake.git
    Commit 5d83d828cacb58ccb7c464e799c85fd2d2a50ccc (tag: 1.44.0)
    Author: Richard Purdie <richard.purdie@linuxfoundation.org>
    Date:   Wed Oct 9 14:10:21 2019 +0100

    bitbake: Update to version 1.44.0

    Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>

Signed-off-by: Cedric Hombourger <Cedric_Hombourger@mentor.com>