patch: Replace custom implementation with OE-core solution
Yes, it was a bad idea to try an own, "simpler" solution. Even the
current iteration still fails to replay patches in case unpack re-ran
from a tarball. In that case, the applied_patches folder was not purged
as I assumed (and as it works for git repos e.g.).

So, let's give up and simply import the OE-core implementation. It only
requires some minor tweaks to adjust to the Isar environment. For that,
we just need to drop PATCHDEPENDENCY, adjust QUILTRCFILE and
PATCH_GIT_USER_* and expand the python module search path.

This now makes quilt another dependency for the build environment. kas
1.1 or newer provides it already.

Based on OE core commit 42047185d55fc878821defec4908b119884d3d07.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>