linux-mainline: fix stripping of .0 from the kernel version
The strip() function was used to strip the trailing .0 from the kernel
version since it is omitted in the upstream archive file names. That
function is actually taking a character set as argument not a substring.
This was causing ARCHIVE_VERSION to expand to "4.19.8" when the user
provided version was "4.19.80". The inline python code was changed to
check if the version ends with ".0" and only then remove the last two
characters.

Signed-off-by: Cedric Hombourger <Cedric_Hombourger@mentor.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>