expand-on-first-boot: run systemd service before local-fs-pre.target
In case that last partition is one of the local filesystems and not the
rootfs we have to be done with the expansion before systemd tries to
mount that filesystem.

Depending on the execution order and speed it was likely to get dropped
into a rescue shell when the resize and the mount happened at the same
time. That problem should be gone with this patch.

Signed-off-by: Henning Schild <henning.schild@siemens.com>