wic-img: allow wic to find bitbake binary
Some code paths/plugins use bitbake libraries. Those are found relative
to the bitbake binary.
Mount bitbake into the buildchroot and rewrite the sudo statement so
that bitbake will be found in PATH.

One example to trigger the original problem would be to use the
"--exclude-path" argument for the "part" command of wic.

Signed-off-by: Henning Schild <henning.schild@siemens.com>