meta-isar: Clean up qemu multiconfigs
Move all the needless duplications from the multiconfigs into the
machine configs, reducing the multiconfigs to MACHINE and DISTRO
settings - with the only exception of qemuamd64-buster-tgz.conf.

Note that this also switches all qemui386 targets to the 686-pae kernel.
So far, only the stretch target used it - for whatever reason.

As the QEMU data is now in the machine config, start_vm also needs to be
updated to pull from there, no longer from any multiconfig.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>