meta: Decouple kernel copy-out from KERNEL_FILE
Apparently, there is no consistent correlation between the kernel image
link in the root directory an the image naming pattern in /boot for the
kernel packages. The RISC-V port uses vmlinuz link names, mips uses
vmlinux. So we can't use KERNEL_FILE as abstraction and should rather
probe for the two possible options.

CC: Cedric Hombourger <Cedric_Hombourger@mentor.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>