bitbake: Update to 1.46.2 release
Update bitbake to the latest 1.46.2 release.
This release is tagged to the below upstream commit ID:
cc11dfa4eb3616547a8a3909f89da0cc4f35dc57

The version in bin/bitbake is not updated to 1.46.2 and still shows 1.46.0.
Please use the above commit ID to validate it is indeed 1.46.2 release.

Signed-off-by: Vijai Kumar K <Vijaikumar_Kanagarajan@mentor.com>