sdk: Add script to relocate SDK
When run inside the unpacked SDK, this script tunes interp and rpath
entry in relevant binaries so that the cross compilation tool can be
called outside of the chroot, irrespective of the host distribution.
Then only "--sysroot /path/to/sdkroot" needs to be passed to the
compiler.

The script also supports restoring the original chroot-mode when invoked
with the --restore-chroot option.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>