sdkchroot: Fix dependency for RISC-V
RISC-V needs our crossbuild-essential-riscv64 package, and that needs to
be added to the dependency list of the sdkchroot, like already
the case for dpkg-base.bbclass.

While we could so far rely on buildchroot-host having installed the
required real dependencies of crossbuild-essential-riscv64 already, this
is not the case for the sdkchroot. So add those dependencies to our
replacement package.

Reported-by Steffen Gerhold <steffen.gerhold@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>