u-boot: add libubootenv
Add the new library libubootenv and remove fw_printenv and fw_setenv
form u-boot-tools as the are now part of the new library.

libubootenv is a library that provides a hardware independent
way to access to U-Boot environment. U-Boot has its default environment
compiled board-dependently and this means that tools to access the environment
are also board specific, too.

libubootenv conflicts with u-boot-tools from Debian 10
as both try to install fw_printenv and fw_sentenv. This conflict is not
part of the control file as it breaks the installation of custom u-boot-tools
from the u-boot-sources.

This patch uses dpkg-gdb to build the package from salsa.debian.org and adds
a fix for https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=967487.

Signed-off-by: Quirin Gylstorff <quirin.gylstorff@siemens.com>
[Jan: add PROVIDES, fix indention, fix de0-nano-soc example]
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>