wic/plugins: Fix boot plugin
When we use --use-uuid along with bootimg plugin, it is observed
that the uuid updated in /etc/fstab is not same as the one the
disk created by bootimg has.

The latest wic[1] creates a UUID for use for these plugins, our ISAR
version of these plugins were not consuming the uuid generated by
wic while generating the partition. Instead a default volume id is
generated which is different from what is written by wic in fstab;
resulting in boot failure.

Fix the ISAR bootimg plugins to use the uuid from wic.

[1] https://github.com/openembedded/openembedded-core/commit/9256b8799495634ee8aee5d16ff71bd6e6e25ed4

Reported-by: Henning Schild <henning.schild@siemens.com>
Signed-off-by: Vijai Kumar K <Vijaikumar_Kanagarajan@mentor.com>
Tested-by: Henning Schild <henning.schild@siemens.com>