meta/isar-bootstrap: allow passing a SCRIPT to debootstrap
Debootstrap uses scripts that usually have the same name as SUITE so we
never set them. In order to install a SUITE that an outdated version of
debootstrap does not (yet) know, we can tell debootstrap which SCRIPT to
use.
In fact many modern SUITEs will use SCRIPTS we might already have, we
just do not have the symlink to them.
i.e. "cosmic" uses the SCRIPT "gutsy" and a debootstrap from "stretch"
does know "gutsy" but not "cosmic"
So we can debootstrap SUITEs from the future, given they use a SCRIPT we
already have.

Signed-off-by: Henning Schild <henning.schild@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>