wic: Warn if an ext filesystem affected by the Y2038 problem is used
We are getting closer and closer to the year 2038 where the 32 bit
time_t overflow will happen. While products (= embedded systems) with an
expected life time of 15 years are still save the situation may change
if your system has to survive the next 20 years.

While ext2 and ext3 file systems are always affected by the time overflow,
let's warn the user if these file systems are still being used.

If ext4 is affected depends on the inode size chosen during file system
creation. At least 256 bytes are necessary to be save. As ext4 is
used very often (and partitions may be small first and extended later)
this might be an issue for many users.

Signed-off-by: Florian Bezdeka <florian.bezdeka@siemens.com>