meta-isar: Add simple example for injecting prebuilt deb packages
While it's definitely nicer and more convenient to pull deb packages
from repos, this may not always be the case. This recipe demonstrates
how to implement such a pattern.

The example embeds the deb as local artifact which is clearly an
anti-pattern. That path was only chosen here because it's hard to find a
stable URL to a prebuilt arch-independent deb file in the wild. The
procedure for creating the deb is also included as README.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>