meta-isar: Fix container configurations
The machine conf was missing the target architecture, and all should in
fact carry that arch (currently only amd64) in their names.

As this avoid building containers for random non-host archs, we can
remove the KFAIL protection from the bullseye variant.

Also fix the copyright headers at this chance and use weak assignments.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>