wic: apply the actual fork changes to our pcbios plugin fork
This aims to be a rather minimal one-commit fork to do isar specific
changes to the forked class. It does things in a way that results in
somehow strange code but pretty clean patching that aims to be easy to
read and mostly about adding lines to get things done. The hope is that
it will make maintaining the fork easier than before.

Signed-off-by: Henning Schild <henning.schild@siemens.com>