wic: apply the actual fork changes to our efi plugin fork
This aims to be a rather minimal one-commit fork to do isar specific
changes to the forked class. It does things in a way that results in
somehow strange code but pretty clean patching that aims to be easy to
read and mostly about adding lines to get things done. The hope is that
it will make maintaining the fork easier than before.

Note that it changes the way we deal with grub, we now do it like OE and
consistent with systemd-boot and syslinux. Meaning we have the
bootloader, its config and kernel+initrd in the /boot partition.
While that breaks kernel updates with "apt-get" it gets things
consistent and less hacky. Such kernel updates anyways only worked with
grub-efi and should rather be solved differently covering all possible
bootloaders.

Signed-off-by: Henning Schild <henning.schild@siemens.com>