base: Move SRC_APT extraction/filtering to dpkg-base class
apt:// SRC_URI entries only make sense for dpkg recipes. So move the
filtering over so that any attempt to use it for other types or recipes
is detected early and that related code is located in the same classes.s

Drop redundant expand=True from getVar at this chance. It's default.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>