rootfs: clean package log files that are not owned by packages
/var/log/* files that are created during build stage and not owned
by any package are not neccessary to be present in rootfs image, as
these log files adds additional size to rootfs image, and also it
create problems for reproducible build functionality.

so this ROOTFS feature 'clean-log-files' should help to clean the log
files when it is enalbed, disable it if we need the log files for
debugging purpose.

ROOTFS_FEATURE += clean-log-files

Signed-off-by: venkata pyla <venkata.pyla@toshiba-tsip.com>