CI: install avocado in venv if host does not have it
Avocado packages are currently not available in Debian. Use virtualenv
and pip for sites that do not provide avocado as part of execution
environment. We expect the execution environment to be Debian-based and
to allow us to sudo (true for kas-isar container).

Use version 88.0 as the latest working one from pip.

Signed-off-by: Henning Schild <henning.schild@siemens.com>
Signed-off-by: Anton Mikanovich <amikan@ilbers.de>