base: remove unneeded "before do_build" task dependencies
This makes dependencies more precise, in preparation for
the introduction of caching.

Currently the dependency graph is overspecified, and do_build has
too many explicit dependencies. What happens then is that even
if a package (the output of do_dpkg_build) can be taken from
cache, there are still dependencies that cause the execution of
do_fetch, do_patch etc.

Signed-off-by: Adriaan Schmidt <adriaan.schmidt@siemens.com>