dpkg: add explicit dependency to isar-apt
Again, make dependencies more precise and explicit.

The tasks do_install_builddeps and do_deploy_deb need
isar-apt, which is usually "just there" because other early tasks
pull this in.

When using caching, one example of what can happen is:
We run with empty builddir and warm cache, and request the
build of a single package. With a cache hit, the only task
that will run is the packages do_deploy_deb. So this needs
an explicit dependency on isar-apt.
Same for install_builddeps and install_imager_deps.

Signed-off-by: Adriaan Schmidt <adriaan.schmidt@siemens.com>