sstate: add caching to rootfs
We cache the output of do_rootfs. Again "sudo tar" to preserve
ownerships and permissions. In addition, it can happen that there are
active mounts while tar'ing the rootfs (there shouldn't be, but I've
seen cases where deb-src is mounted). Unfortunately --one-file-system
does not stop at bind mounts, so we use a trick with a temporary mount
to ensure that we don't package more than we want.

Special care needs to be taken with image and initramfs, which are
MACHINE specific. So they adapt sstate manifest location to ensure
that there is no false sharing.

Signed-off-by: Adriaan Schmidt <adriaan.schmidt@siemens.com>