refactor: inject ccache environment similar to proxy env
Instead of configuring ccache in the build shell scripts,
we configure it by preparing the environment of the calling task.

The tasks invoke the build scripts using sudo.
By that, we have to modify the PATH variable inside sudo as
sudo itself clears it and sets it to the value of secure_path.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>