vm-img: use two different ovf templates for vmware/virtualbox
Both vmware and virtualbox have their own dialects of "ovf", there is no
common way to describe things. That combined template was trying by
mixing both dialects into one file. But the result caused parsing error
on recent vmware.

Instead of trying one size fits all, this patch introduces two
templates. Both derived from ovf exports done on recent version of
vmware and virtualbox. Because of the vmdk sub-format it was anyways not
possible to build one ova image that would work for both VMMs.

We also now set the uppercase version of OVA_FIRMWARE, which is needed
for virtualbox. Before it was called OVA_FIRMWARE_VIRTUALBOX but was not
set.

Signed-off-by: Henning Schild <henning.schild@siemens.com>