Revert "buildchroot: Do not build cross when there are no arch-specific outputs"
This reverts commit 563986703d9a0405c70af9b50ecedede2ac59cbd.

The reverted patch made a shortcut to simplify the handling of cross-build
dependencies of architecture:all packages.
This is not valid for the following reasons:

1. Just scanning the control file for 'all' is not sufficient as
a source package might generate both arch specific and :all packages.

2. This breaks architecture specific transitive build dependencies.
These mainly apply to correctly packaged python packages that support
cross building and use setuptools. By that, the transitive dependency
libpython3.9-minimal:<target-arch> is not attracted but instead the version
for the host, leading to very hard to debug build-time issues.

In case packages do not handle their cross-build dependencies correctly, we
should not try to work around in ISAR, but better enforce upstream patches.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>