fix(sstate): don't fail if there are no packages
The code to put generated deb packages into the sstate cache
currently uses the pattern
  test <condition> && do_stuff
which, if the condition is not met, not only skips do_stuff,
but also returns failure.

The consequence is that in cases where there are no packages
found in ${S}/../*.deb, the sstate caching fails completely.

This changes that pattern to use an explicit "if" instead.

Signed-off-by: Adriaan Schmidt <adriaan.schmidt@siemens.com>