u-boot-script: Allow OVERLAY to be a u-boot env variable as well
This allows to define the list of overlays also during boot-time,
retrieving it from a u-boot env variable. All we need to do for that is
adding a loop around loading and applying entries in the generated
script. That will interpret an entry during boot-time as list and
iterate over all entries.

In order to use this feature, the wks entry for a rootfs-u-boot
partition needs to be extended by

--sourceparams "overlays=\"\$\{my_overlays\}\""

Quoting and escaping is important to propagate ${my_overlays} into the
generated script.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>