bootstrap: do not leave 50isar.conf in final rootfss
This commits makes sure we do not leave a file around which is really
only meant for building. In case people will want to "apt-get" later
they might have their own ideas on how to deal with Recommends and
Suggests. And if so they should write a package to configure apt for
their images.

So it cleans up a file not owned by any package, and unifies the way apt
configuration for the build is done.

Signed-off-by: Henning Schild <henning.schild@siemens.com>