CI: pin down the kas container version we use in CI
That is a very good idea anyhow, even though it will need regular
version bumping and cause some work.

But right now we are doing it to work around a problem in debian 11.

In Debian >= bullseye 'dpkg-deb --show --showformat' is currently broken
in case of requesting virtual fields like 'source:<FieldName>'.
This makes function 'debsrc_download' broken, so build fails when
ISAR_USE_CACHED_BASE_REPO is enabled in local.conf

The regression came with the fix for bug #972580 in Debian.
The issue is reported to Debian bugtracker as bug #1004372.

Signed-off-by: Henning Schild <henning.schild@siemens.com>