linux-module: remove double quotes around AUTOLOAD
If the AUTOLOAD variable has more than 1 entry the for-loop in
do_prepare_build() doesn't expand and therefor result in a
one-line entry into the postinst file. This in turn gives the
wrong format for the /etc/modules (all module names are on the
same line) and therefor fails to load at startup.

Signed-off-by: Daniel Machon <dama@universal-robots.com>
Signed-off-by: Torsten Bonde Christensen <tbch@universal-robots.com>