meta: add quality check helper to image class
Content of rootfs should ideally all be coming from packages and their
hooks. POSTPROCESSing and custom tasks should be handled with a lot of
care and avoided where possible.

This commit introduces a quality check task to help devs not
accidentially abuse POSTPROCESS, while allowing them to add files to an
ignore list.

Signed-off-by: Henning Schild <henning.schild@siemens.com>