image-locales-extension: Do localepurge configuration after installation
The current seqence of copying the localepurge configuration first and
then  installing localepurge package is causing the debconf cache entry,
that is because the package postinstall script assume the configuration
is already present and checks for configuration change and updates the
debconf cache entry.

With the above sequence of steps it is causing unnecessary debconf cache
entry and which is causing non-reproducible system image creation.

To avoid this, the sequence is altered to install the package first and
then configure necessary locales to remove.

Signed-off-by: venkata pyla <venkata.pyla@toshiba-tsip.com>