classes/image-postproc-extension: Remove /etc/machine-id
In a read-only system the machine id should be deleted or
set to `unitialized\n`[1].

Systemd will generate a new machine-id during the first boot.
In the case of a read-only root file system Systemd generates a mount point with
the machine id. If an overlay for /etc is used this creates a mount conflict.
To avoid the conflict between the overlay filesystem and systemd /etc/machine-id
mount point deleted the file /etc/machine-id.

This partly reverts 98d6a72d3064. The man page[2] was updated for Debian Bullseye to
allow a missing machine id. For Debian Bullseye it reads:

```
For operating system images which are created once and used on multiple machines,
for example for containers or in the cloud, /etc/machine-id should be either missing
or an empty file in the generic file system image
```

[1]: https://systemd.io/BUILDING_IMAGES/
[2]: https://manpages.debian.org/bullseye/systemd/machine-id.5.en.html

Signed-off-by: Quirin Gylstorff <quirin.gylstorff@siemens.com>