expand-on-first-boot: Add support for devicemapper
When the rootfs is inside a devicemapper volume, e.g. dm-verity, the
script so far fails to identify the boot device. This adds at least
basic support for that case (single level, single slave). Once there is
a good example for more complex setups, this could be taken into account
as well.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>