rpi: make example image support Y2038
In the build we actually receive a warning
WARNING: mc:rpi-arm-bullseye:isar-image-base-1.0-r0 do_wic_image:
WARNING: The ext2 filesystem mounted at / has no Y2038 support.

This commit switches over to ext4 and appends the mke2fs args we need to
handle y2038, with it the warning is gone. And people that might take
that wks will not see a bad example.

Tested on a "Raspberry Pi 3 Model B Rev 1.2"

Signed-off-by: Henning Schild <henning.schild@siemens.com>