bitbake-diffsigs: break on first dependent task difference
Currently compare_sigfiles() recursively calculates differences on all
dependent tasks with changed hashes. This is done in arbitrary order, and
only the last of those results is returned while everything else is discarded.

This changes that to instead return the first difference and not calculate
any more, which significantly speeds up diffs of tasks with many dependencies.

backported from upstream:
https://git.openembedded.org/bitbake/commit/?id=8ed7722865d2dcfda1697aaf4e12b828934bf527

Signed-off-by: Adriaan Schmidt <adriaan.schmidt@siemens.com>