utils: Ensure shell function failure in python logging is correct
If a python function exec_func() calls a shell task, the logging wasn't working
correctly in all cases since the exception was turned into a BBHandledException()
and the logfile piece was lost which is handled at the top task level.

The easiest way to avoid this is to allow the ExecutionError exceptions to
be raised to a higher level, we don't need the traceback for them.

Backported from Bitbake 2.0 upstream branch:
https://git.openembedded.org/bitbake/commit/?id=7cae11f558f9ff5fd05ef23b789aaef92fb5a327

Signed-off-by: Anton Mikanovich <amikan@ilbers.de>