buildchroot: Install ccache into the buildchroot only when enabled
This fixes a corner case where the buildchroot is not directly fetched
from Debian upstream repos but from something like a (manually
maintained) snapshot. There is no guarantee that the ccache packet is
available inside the buildchroot packet feed.

As soon as ccache is enabled by setting "USE_CCACHE" to "1" the packet
will be installed into the buildchroot.

Signed-off-by: Florian Bezdeka <florian.bezdeka@siemens.com>