rootfs: clean debconf cache files
Cleaning the debconf cache files for the purpose of generating
reproducible system image.

The debconf cache database sometime generates with non-reproducible
contents because of some packages behaviour (e.g: localepurge), reported
this to package bug tracking system[1], but because this problem is not
with one particular package and can happen with other packages also,
this should be solved in common place to address reproducibility
problem, discussing with Reproducible-builds community[2] to find some
common solution for this problem.

[1] https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=1009639
[2] https://lists.reproducible-builds.org/pipermail/rb-general/2022-April/002546.html

Signed-off-by: venkata pyla <venkata.pyla@toshiba-tsip.com>
Suggested-by: Henning Schild <henning.schild@siemens.com>