linux-custom: Fix error handling of kernel build
Error was widely broken because set -e was not consistently set and,
where it was, it had no effect on functions as the do_${target}
invocation put them in a pipe which causes -e to be ignored.

This sets -e in the common script already and avoids the now unneeded
"|| exit ${?}".

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>